+++ test/integration/Integration/Miscellaneous.hs
@@ -224,10 +224,13 @@ lineTruncDetect = testPropertyNamed desc "lineTruncDetect"
     logs <- liftIO $ readIORef logsRef
     logs === []
   where
-    desc = "cmdLogLineTrunc reads 'detect' string from toml"
+    desc = "lineTrunc reads 'detect' string from toml"
     args = ["-c", getIntConfig "misc", "cmd1"]
 
-    expected = [#coreConfig % #consoleLogging % #lineTrunc % _Just ^?=@ Just 87]
+    expected =
+      [ #coreConfig % #consoleLogging % #lineTrunc % _Just ^?=@ Just 87,
+        #coreConfig % #fileLogging %? #lineTrunc % _Just ^?=@ Just 87
+      ]
 
 testFileSizeModeNothing :: TestTree
 testFileSizeModeNothing = testPropertyNamed desc "testFileSizeModeNothing"